Added more ClientManagerTest to test updating things in the DB
These tests are not finished yet because I have a doubt regarding the
update methods, at this point they are only updating the DB but not
setting the Java object. Is the responsibility of the update method to
also set the java object or not? -> comment this to the team