Established the access to the DoctorMenu through the MainMenu
I have established the connection between DoctorMenu and MainMenu via
constructor of DoctorMenu. Added a new constructor in Order so it
receives a client as parameter and internally creates an empty cart
(empty ArrayList of ProductOders).
There are some methods in DoctorMenu that are not finished yet and need
to be refactored