Solved the database problem for good (I hope)
I corrected the methods that import products and suppliers into the
database from CSV files, checking whether they have already been
inserted or not, and acting in one way or another depending on the
result of the existence check. I also changes things in the menus and
oversaw some warning messages.